Added service to user controller and changet the type of all id-s in controllers to Long from integer.

+++ src/main/java/com/ISS/Booking_iss_tim21/controller/AccommodationReviewController.java
@@ -35,7 +35,7 @@ public class AccommodationReviewController {
     }
 
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<Void> deleteAccommodationReview(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteAccommodationReview(@PathVariable Long id) {
 
         AccommodationReview review = service.findOne(id);
